+++ app/services/data_product/models/search_data_products.py
@@ -4,14 +4,15 @@ from typing import Union, Literal, List
 
 class SearchDataProductsRequest(BaseModel):
     product_search_query: Union[Literal["*"], str] = Field(
-        description='The search query to search for data products. If the user wants to search for data products with a specific name, this is the name to search for. If user wants to search for all data products, this value should be "*".'
+        description='The search query to search for data products and data product drafts. If the user wants to search for data products or drafts with a specific name, this is the name to search for. If user wants to search for all data products or drafts, this value should be "*".'
     )
     search_filter_type: Literal["None", "Domain"] = Field(
         default="None",
         description="Specify what to filter by. It can be one of the following: None, Domain. If the user wants to filter by domain, then this value should be Domain otherwise None.",
     )
     search_filter_value: str = Field(
-        description="The value to filter by. For example, if search_filter_type is Domain, then this is the domain name to filter by."
+        default="None",
+        description="The value to filter by. For example, if search_filter_type is Domain, then this is the domain name to filter by. Use 'None' when search_filter_type is 'None'."
     )